device/requirements.txt
RPi.GPIO
adafruit-circuitpython-dht
adafruit-circuitpython-ads1x15
smbus2